beauty-ai-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pandas==2.1.3
Pillow==10.1.0
groq==0.11.0
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0